24/04
- Player data between scenes
- Map and Key Pick Ups
- Better Skull Level Changes

+++ TreasureICA/Assets/My Assets/Scripts/PickUpScripts/ApplePickUp.cs
@@ -6,7 +6,7 @@ public class ApplePickUp : MonoBehaviour {
     [SerializeField]
     private PlayerHealth playerHealth;
 
-    private float increaseHealthAmount = 20.0f;
+    private int increaseHealthAmount = 20;
 
 	// Use this for initialization
 	void Start ()
